refactor(Auth): Auth
Auth 폴더를 common 폴더 하위로 이동하였습니다.
