+++ src/main/java/com/space/server/common/auth/service/implementation/AdditionalInfoUpdater.java
@@ -1,7 +1,7 @@
-package com.space.server.auth.service.implementation;
+package com.space.server.common.auth.service.implementation;
 
-import com.space.server.auth.domain.repository.RefreshRepository;
-import com.space.server.auth.presentation.dto.request.AdditionalInfoRequest;
+import com.space.server.common.auth.domain.repository.RefreshRepository;
+import com.space.server.common.auth.presentation.dto.request.AdditionalInfoRequest;
 import com.space.server.common.jwt.exception.InvalidTokenException;
 import com.space.server.common.jwt.exception.RefreshTokenNotFoundException;
 import com.space.server.common.jwt.util.JwtUtil;
